Fix for Playlist Manager UI not appearing on Load
UI was not appearing/injecting on load (if loaded page was a playlist). And also navigating to a non-playlist page and back to a playlist wouldn't inject the UI. but when you navigate to another page the UI would popup. Fixed all of these issues by adding a permanent history listener so that we are always detecting the current page at any point and then applying functionality/logic from there.